requirements.txt
python-telegram-bot==20.8
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
playwright==1.44.0
beautifulsoup4==4.12.3
requests==2.31.0
